Rename schema file
