+++ services/transaction/shared/schema/multisigTransaction.js
